adress comment to move test into new file
